Moved central knowledge to DB as they share same operations
