+++ reguMatch/api/tools/database/models.py
@@ -1,6 +1,7 @@
 from pydantic import BaseModel, Field, field_validator
 from datetime import datetime
 from typing import List, Optional, Dict
+from enum import Enum
 
 class CountryInformation(BaseModel):
     country_name: str = Field(
@@ -170,3 +171,50 @@ class WhiteListQueryResponse(BaseModel):
         description="The document that was retrieved from mongoDB."
     )
     error: Optional[str] = Field(None, description="An error message if the operation failed.")
+
+class RequirementType(str, Enum):
+    REGULATION = "REGULATION" # Gov stuff
+    COMPLIANCE = "COMPLIANCE" # Non-gov stuff
+
+class RegulationAndComplianceBusinessCriteria(BaseModel):
+    business_size: str = Field(..., description="Which business sizes this applies to. e.g. ['Large', 'Enterprise']. If None, applies to all sizes.")
+    business_type: str = Field(..., description="Which business types this applies to. e.g. ['Pty Ltd', 'Public Company']. If None, applies to all types.")
+    business_sector: str = Field(..., description="The sector that this applies to. e.g. ['Commercial Fishing']. If None, applies to all activities.")
+    custom_criteria: Dict[str, str] = Field(..., description="Custom criteria specific to the industry. e.g. {'vessel_length_meters': '>24', 'crew_size': '>20'}")
+
+class BaseRegulationAndComplianceRequirements(BaseModel):
+    location_information: CountryInformation = Field(..., description="The location where this regulation and compliance requirements apply.")
+    industry_information: CategoryInformation = Field(..., description="The industry where this regulation and compliance requirements apply.")
+    business_criteria: RegulationAndComplianceBusinessCriteria = Field(..., description="The business criteria where this regulation and compliance requirements apply.")
+    requirement_type: RequirementType = Field(..., description="The type of regulation and compliance requirements. e.g. REGULATION or COMPLIANCE")
+
+class RequiredLicenses(BaseModel):
+    license_name: str = Field(..., description="The name of the license.")
+    license_description: str = Field(..., description="The description of the license.")
+
+class RequiredFields(BaseModel):
+    field_name: str = Field(..., description="The name of the field.")
+    field_description: str = Field(..., description="The description of the field.")
+
+class RegulationNode(BaseRegulationAndComplianceRequirements):
+    regulation_name: str = Field(..., description="The name of the regulation.")
+    regulation_description: str = Field(..., description="The description of the regulation.")
+    regulation_webpage: str = Field(..., description="The webpage of the regulation.")
+    regulation_pdf_url: Optional[str] = Field(..., description="The PDF URL of the regulation.")
+
+    required_fields: Optional[List[RequiredFields]] = Field(..., description="The fields required for this regulation.")
+    required_licenses: Optional[List[RequiredLicenses]] = Field(None, description="The licenses required for this regulation.")
+
+    regulatory_body: str = Field(..., description="The name of the regulatory body.")
+    regulatory_body_description: str = Field(..., description="The description of the regulatory body.")
+
+
+
+class ComplianceNode(BaseRegulationAndComplianceRequirements):
+    compliance_name: str = Field(..., description="The name of the compliance.")
+    compliance_description: str = Field(..., description="The description of the compliance.")
+    compliance_webpage: str = Field(..., description="The webpage of the compliance.")
+    compliance_pdf_url: Optional[str] = Field(..., description="The PDF URL of the compliance.")
+
+    required_fields: Optional[List[RequiredFields]] = Field(..., description="The fields required for this compliance.")
+    required_licenses: Optional[List[RequiredLicenses]] = Field(None, description="The licenses required for this compliance.")
