+++ reguMatch/api/tools/database/service.py
@@ -1,7 +1,7 @@
 from fastmcp import FastMCP
 from contextlib import asynccontextmanager
 import json
-from api.tools.database.models import WhiteListURLData, WhiteListQueryParameters
+from api.tools.database.models import WhiteListURLData, WhiteListQueryParameters, ComplianceNode, RegulationNode
 from api.database.mongo_client import MongoDBClient
 
 from api.database.mongo_queries import (
@@ -178,3 +178,25 @@ async def get_whitelist_collection() -> str:
             },
             indent=2,
         )
+
+@mcp.tool(name="add_compliance_node",
+          description="Add a new compliance node to the central knowledge graph",)
+async def add_compliance_node(compliance_node: ComplianceNode) -> str:
+    print(json.dumps(compliance_node.model_dump(), indent=2))
+    return "success"
+
+@mcp.tool(name="add_regulation_node",
+          description="Add a new regulation node to the central knowledge graph",)
+async def add_regulation_node(regulation_node: RegulationNode) -> str:
+    print(json.dumps(regulation_node.model_dump(), indent=2))
+    return "success"
+
+@mcp.tool(name="get_regulation_node",
+          description="Get a regulation node from the central knowledge graph",)
+async def get_regulation_node() -> str:
+    pass
+
+@mcp.tool(name="get_compliance_node",
+          description="Get a compliance node from the central knowledge graph",)
+async def get_compliance_node() -> str:
+    pass
